relocate load_sample to yt.loaders
