+++ yt/loaders.py
@@ -9,8 +9,9 @@ import os
 
 import numpy as np
 
+import yt.utilities.sample_data as sd
 from yt.config import ytcfg
-from yt.funcs import ensure_list, issue_deprecation_warning
+from yt.funcs import ensure_list, issue_deprecation_warning, mylog
 from yt.utilities.decompose import decompose_array, get_psize
 from yt.utilities.exceptions import (
     YTAmbiguousDataType,
@@ -20,13 +21,11 @@ from yt.utilities.exceptions import (
 )
 from yt.utilities.hierarchy_inspection import find_lowest_subclasses
 from yt.utilities.lib.misc_utilities import get_box_grids_level
-from yt.utilities.logger import ytLogger as mylog
+from yt.utilities.on_demand_imports import _pooch as pch
 from yt.utilities.parameter_file_storage import (
     output_type_registry,
     simulation_time_series_registry,
 )
-
-from .utilities.load_sample import load_sample  # noqa F401
 
 # --- Loaders for known data formats ---
 
@@ -1279,3 +1278,140 @@ def load_unstructured_mesh(
     sds.default_field = [f for f in sds.field_list if f[0] == "connect1"][-1]
 
     return sds
+
+
+# --- Loader for yt sample datasets ---
+# This utility will check to see if sample data exists on disc.
+# If not, it will download it.
+
+
+def load_sample(name=None, specific_file=None, pbar=True):
+    """
+    Load sample data with yt. Simple wrapper around yt.load to include fetching
+    data with pooch.
+
+    Parameters
+    ----------
+    name : str or None
+        The name of the sample data to load. This is generally the name of the
+        folder of the dataset. For IsolatedGalaxy, the name would be
+        `IsolatedGalaxy`.  If `None` is supplied, the return value
+        will be a list of all known datasets (by name).
+
+    specific_file : str, optional
+        optional argument -- the name of the file to load that is located
+        within sample dataset of `name`. For the dataset `enzo_cosmology_plus`,
+        which has a number of timesteps available, one may wish to choose
+        DD0003. The file specifically would be
+        `enzo_cosmology_plus/DD0003/DD0003`, and the argument passed to this
+        variable would be `DD0003/DD0003`
+
+    pbar: bool
+        display a progress bar
+
+    """
+
+    fido = sd.Fido()
+    if name is None:
+        keys = []
+        for key in fido._registry:
+            for ext in sd._extensions_to_strip:
+                if key.endswith(ext):
+                    key = key[: -len(ext)]
+            keys.append(key)
+        return keys
+
+    base_path = fido.fido.path
+    fileext, name, extension = _validate_sampledata_name(name)
+
+    downloader = None
+    if pbar:
+        try:
+            import tqdm  # noqa: F401
+
+            downloader = pch.pooch.HTTPDownloader(progressbar=True)
+        except ImportError:
+            mylog.warning("tqdm is not installed, progress bar can not be displayed.")
+
+    if extension == "h5":
+        processor = pch.pooch.Untar()
+    else:
+        # we are going to assume most files that exist on the hub are
+        # compressed in .tar folders. Some may not.
+        processor = None
+
+    fname = fido.fido.fetch(name, processor=processor, downloader=downloader)
+
+    # The `folder_path` variable is used here to notify the user where the
+    # files have been unpacked to. However, we can't assume this is reliable
+    # because in some cases the common path will overlap with the `load_name`
+    # variable of the file.
+    folder_path = os.path.commonprefix(fname)
+    mylog.info("Files located at %s", folder_path)
+
+    # Location of the file to load automatically, registered in the Fido class
+    info = fido[fileext]
+    file_lookup = info["load_name"]
+    optional_args = info["load_kwargs"]
+
+    if specific_file is None:
+        # right now work on loading only untarred files. build out h5 later
+        mylog.info("Default to loading %s for %s dataset", file_lookup, name)
+        loaded_file = os.path.join(base_path, "%s.untar" % fileext, name, file_lookup)
+    else:
+        mylog.info("Loading %s for %s dataset", specific_file, name)
+        loaded_file = os.path.join(base_path, "%s.untar" % fileext, name, specific_file)
+
+    return load(loaded_file, **optional_args)
+
+
+def _validate_sampledata_name(name):
+    """
+    format name of sample data passed to function, accepts a named string
+    argument and parses it to determine the sample data name, what type of
+    extension it has, or other relevant information.
+
+    returns
+    -------
+    fileext : str
+        The name of the sample data, with the file extension
+        example: "IsolatedGalaxy.tar.gz"
+    basename : str
+        The name of the sample data, without the file extension
+        example: "IsolatedGalaxy"
+    extension : str
+        name of extension of remote sample data
+        example: "h5" or "tar"
+    """
+
+    if not isinstance(name, str):
+        mylog.error("The argument %s passed to load_sample() is not a string.", name)
+
+    # now get the extension if it exists
+    base, ext = os.path.splitext(name)
+    if ext == "":
+        # Right now we are assuming that any name passed without an explicit
+        # extension is packed in a tarball. This logic can be modified later to
+        # be more flexible.
+        fileext = "%s.tar.gz" % name
+        basename = name
+        extension = "tar"
+    elif ext == ".gz":
+        fileext = name
+        basename = os.path.splitext(base)[0]
+        extension = "tar"
+    elif ext in [".h5", ".hdf5"]:
+        fileext = name
+        basename = base
+        extension = "h5"
+    else:
+        mylog.info(
+            """extension of %s for dataset %s is unexpected. the `load_data`
+            function  may not work as expected""",
+            ext,
+            name,
+        )
+        extension = ext
+        fileext = name
+        basename = base
+    return fileext, basename, extension
